Move data-modelling classes into data module
